try fix imports

+++ final_task/additional_func/pycalc_common.py
@@ -1,5 +1,5 @@
 import re
-from additional_func import calculate_math_parameters, calculate_expression, parse_brackets
+from .additional_func import calculate_math_parameters, calculate_expression, parse_brackets
 import math
 
 
@@ -16,4 +16,4 @@ def pcalc(value):
         result = "".join(calculate_math_parameters.calc_math_parameter(value))
     result = parse_brackets.parse_brackets(result)
     result = calculate_expression.calculate(result)
-    return float(result)
+    return int(result)
